Test WorldRecordRepository
